Вынес компонент Nutrient в отдельный файл
