+++ src/components/IngredientDetails/IngredientDetails.module.css
@@ -10,9 +10,3 @@
   text-align: center;
 }
 
-.nutrient {
-  display: flex;
-  justify-content: center;
-  flex-direction: column;
-  align-items: center;
-}
