+++ src/components/IngredientDetails/Nutrient/Nutrient.module.css
@@ -1,0 +1,6 @@
+.nutrient {
+  display: flex;
+  justify-content: center;
+  flex-direction: column;
+  align-items: center;
+}
