+++ src/components/IngredientDetails/index.js
@@ -1,3 +1,4 @@
 import IngredientDetails from "./IngredientDetails";
+import Nutrient from "./Nutrient/Nutrient";
 
-export { IngredientDetails };
+export { IngredientDetails, Nutrient };
